+++ _pyproject.toml
@@ -24,10 +24,10 @@ dev = [
     "mypy",
     "pytest",
     "pytest-cov",
-    "cython",
     "setuptools==71.1.0",
     "numpy",
     "gurobipy",           # used on CI to confirm equality of results
+    "pyscipopt",           # used on CI to confirm equality of results
 ]
 docs = ["sphinx", "sphinx_rtd_theme", "sphinx-autodoc-typehints"]
